返ってくるのはint64でした

+++ api/domain/user.go
@@ -4,7 +4,7 @@ import "github.com/jinzhu/gorm"
 
 type User struct {
 	gorm.Model
-	TootsCount      uint
-	FollowsCount    uint
-	FollowersSecret uint
+	TootsCount      int64
+	FollowsCount    int64
+	FollowersSecret int64
 }
